requirements.txt.txt
streamlit
pdf2image
pytesseract
python-docx